requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.5
django-taggit==6.1.0
djangorestframework==3.15.2
idna==3.10
pillow==11.1.0
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.32.3
sqlparse==0.5.3
text-unidecode==1.3
urllib3==2.3.0
